+++ public/services/routes.js
@@ -1,6 +1,53 @@
 // Require routes
 let routes = require('ui/routes');
 
+// Don't know why, but all these dependencies must be here...
+
+// Added dependencies (from Kibana module)
+import 'ui/pager';
+import 'ui/typeahead';
+
+// Added from its index.js
+import 'plugins/kibana/discover/saved_searches/saved_searches';
+import 'plugins/kibana/discover/directives/no_results';
+import 'plugins/kibana/discover/directives/timechart';
+import 'ui/collapsible_sidebar';
+import 'plugins/kibana/discover/components/field_chooser/field_chooser';
+import 'plugins/kibana/discover/controllers/discover';
+import 'plugins/kibana/discover/styles/main.less';
+import 'ui/doc_table/components/table_row';
+import _ from 'lodash';
+import angular from 'angular';
+import { getSort } from 'ui/doc_table/lib/get_sort';
+import * as columnActions from 'ui/doc_table/actions/columns';
+import dateMath from '@elastic/datemath';
+import 'ui/doc_table';
+import 'ui/visualize';
+import 'ui/notify';
+import 'ui/fixed_scroll';
+import 'ui/directives/validate_json';
+import 'ui/filters/moment';
+import 'ui/courier';
+import 'ui/index_patterns';
+import 'ui/state_management/app_state';
+import 'ui/timefilter';
+import 'ui/share';
+import 'ui/query_bar';
+import { VisProvider } from 'ui/vis';
+import { BasicResponseHandlerProvider } from 'ui/vis/response_handlers/basic';
+import { DocTitleProvider } from 'ui/doc_title';
+import PluginsKibanaDiscoverHitSortFnProvider from 'plugins/kibana/discover/_hit_sort_fn';
+import { FilterBarQueryFilterProvider } from 'ui/filter_bar/query_filter';
+import { AggTypesBucketsIntervalOptionsProvider } from 'ui/agg_types/buckets/_interval_options';
+import { stateMonitorFactory } from 'ui/state_management/state_monitor_factory';
+import uiRoutes from 'ui/routes';
+import { uiModules } from 'ui/modules';
+import indexTemplate from 'plugins/kibana/discover/index.html';
+import { StateProvider } from 'ui/state_management/state';
+import { migrateLegacyQuery } from 'ui/utils/migrateLegacyQuery';
+import { QueryManagerProvider } from 'ui/query_manager';
+import { SavedObjectsClientProvider } from 'ui/saved_objects';
+
 //Installation wizard
 const settingsWizard = ($location, $q, Notifier, testAPI, appState) => {
     const notify = new Notifier();
@@ -41,6 +88,50 @@ const goToKibana = ($location, $window) => {
     $window.location.href = $location.absUrl().replace('/wazuh#', '/kibana#');
 };
 
+const getIp =  (Promise, courier, config, $location, Private) => {
+    console.log("Resolving my IP");
+    const State = Private(StateProvider);
+    const savedObjectsClient = Private(SavedObjectsClientProvider);
+
+    return savedObjectsClient.find({
+        type: 'index-pattern',
+        fields: ['title'],
+        perPage: 10000
+    })
+    .then(({ savedObjects }) => {
+        /**
+         *  In making the indexPattern modifiable it was placed in appState. Unfortunately,
+         *  the load order of AppState conflicts with the load order of many other things
+         *  so in order to get the name of the index we should use, and to switch to the
+         *  default if necessary, we parse the appState with a temporary State object and
+         *  then destroy it immediatly after we're done
+         *
+         *  @type {State}
+         */
+        const state = new State('_a', {});
+
+        const specified = !!state.index;
+        const exists = _.findIndex(savedObjects, o => o.id === state.index) > -1;
+        const id = exists ? state.index : config.get('defaultIndex');
+        state.destroy();
+
+        return Promise.props({
+            list: savedObjects,
+            loaded: courier.indexPatterns.get(id),
+            stateVal: state.index,
+            stateValFound: specified && exists
+        });
+    });
+};
+
+const getSavedSearch = (courier, savedSearches, $route) => {
+    return savedSearches.get($route.current.params.id)
+    .catch(courier.redirectWhenMissing({
+        'search': '/discover',
+        'index-pattern': '/management/kibana/objects/savedSearches/' + $route.current.params.id
+    }));
+};
+
 //Routes
 routes.enable();
 routes
@@ -59,7 +150,9 @@ routes
     .when('/overview/', {
         template: require('plugins/wazuh/templates/overview.jade'),
         resolve: {
-            "checkAPI": settingsWizard
+            "checkAPI": settingsWizard,
+            "ip": getIp,
+            "savedSearch": getSavedSearch
         }
     })
     .when('/dashboard/:select?', {
@@ -68,10 +161,12 @@ routes
             "checkAPI": settingsWizard
         }
     })
-    .when('/discover/', {
+    .when('/wazuh-discover/', {
         template: require('plugins/wazuh/templates/discover.jade'),
         resolve: {
-            "checkAPI": settingsWizard
+            "checkAPI": settingsWizard,
+            "ip": getIp,
+            "savedSearch": getSavedSearch
         }
     })
     .when('/settings/:tab?/', {
