+++ public/templates/directives/discover-template.html
@@ -1,0 +1,153 @@
+<discover-app class="app-container">
+
+  <!-- Local nav. -->
+  <kbn-top-nav name="discover" config="topNavMenu">
+    <!-- Transcluded elements. -->
+    <div data-transclude-slots>
+      <!-- Breadcrumbs. -->
+      <div data-transclude-slot="topLeftCorner" class="kuiLocalBreadcrumbs">
+        <h1 tabindex="0" id="kui_local_breadcrumb" class="kuiLocalBreadcrumb">
+          <span ng-show="opts.savedSearch.id" class="kuiLocalBreadcrumb__emphasis">
+            <span data-test-subj="discoverCurrentQuery" ng-bind="opts.savedSearch.lastSavedTitle"></span>
+            <i id="reload_saved_search" tabindex="0" aria-label="Reload Saved Search" tooltip="Reload Saved Search" ng-click="resetQuery();" class="fa fa-undo small"></i>&nbsp;
+          </span>
+          <span data-test-subj="discoverQueryHits" class="kuiLocalBreadcrumb__emphasis">{{(hits || 0) | number:0}}</span>
+          <ng-pluralize count="hits" when="{'1':'hit', 'other':'hits'}"></ng-pluralize>
+        </h1>
+      </div>
+
+      <!-- Search. -->
+      <div data-transclude-slot="bottomRow" class="fullWidth">
+        <query-bar
+          query="state.query"
+          app-name="'discover'"
+          on-submit="updateQueryAndFetch($query)"
+        >
+        </query-bar>
+      </div>
+    </div>
+  </kbn-top-nav>
+
+  <main class="container-fluid">
+    <div class="row">
+      <filter-bar
+        state="state"
+        index-patterns="[indexPattern]"
+        ng-if="state.query.language === 'lucene'"
+      ></filter-bar>
+    </div>
+    <div class="row">
+      <div class="col-md-2 sidebar-container collapsible-sidebar" id="discover-sidebar">
+        <disc-field-chooser
+          columns="state.columns"
+          hits="rows"
+          field-counts="fieldCounts"
+          index-pattern="searchSource.get('index')"
+          index-pattern-list="opts.indexPatternList"
+          state="state"
+          on-add-field="addColumn"
+          on-add-filter="filterQuery"
+          on-remove-field="removeColumn"
+        >
+        </disc-field-chooser>
+      </div>
+
+      <div class="discover-wrapper col-md-10">
+        <div class="discover-content">
+          <discover-no-results ng-show="resultState === 'none'"></discover-no-results>
+
+          <!-- loading -->
+          <div ng-show="resultState === 'loading'">
+            <div class="discover-overlay">
+              <h2>Searching</h2>
+              <div ng-show="fetchStatus">{{fetchStatus.complete}}/{{fetchStatus.total}}</div>
+            </div>
+          </div>
+
+          <!-- result -->
+          <div class="results" ng-show="resultState === 'ready'">
+            <button
+              class="kuiButton kuiButton--basic kuiButton--iconText discover-skip-button"
+              ng-click="showAllRows(); scrollToBottom()"
+            >
+              <span class="kuiButton__inner">
+                <span aria-hidden="true" class="kuiButton__icon kuiIcon fa-chevron-down"></span>
+                <span>Skip to bottom</span>
+              </span>
+            </button>
+
+            <div class="discover-timechart" ng-if="opts.timefield">
+              <header>
+                <center class="small">
+                  <span tooltip="To change the time, click the clock icon in the navigation bar">{{timeRange.from | moment}} - {{timeRange.to | moment}}</span>
+
+                  &mdash;
+
+                  <span class="results-interval form-inline">
+                    <select
+                      class="form-control"
+                      ng-model="state.interval"
+                      ng-options="interval.val as interval.display for interval in intervalOptions | filter: intervalEnabled"
+                      ng-blur="toggleInterval()"
+                      data-test-subj="discoverIntervalSelect"
+                      >
+                    </select>
+                    <span ng-show="bucketInterval.scaled">
+                      <kbn-info
+                        placement="right"
+                        class="kuiIcon--info"
+                        info="This interval creates {{ bucketInterval.scale > 1 ? 'buckets that are too large' : 'too many buckets' }} to show in the selected time range, so it has been scaled to {{ bucketInterval.description }}">
+                      </kbn-info>
+                      Scaled to {{ bucketInterval.description }}
+                    </span>
+                  </span>
+                </center>
+
+              </header>
+
+              <visualization
+                 ng-if="vis && rows.length != 0"
+                 vis="vis"
+                 ui-state="uiState"
+                 vis-data="visData"
+                 show-spy-panel="true"
+                 search-source="searchSource"
+                 style="height: 200px"
+                >
+              </visualization>
+            </div>
+
+            <div class="discover-table" fixed-scroll>
+              <doc-table
+                hits="rows"
+                index-pattern="indexPattern"
+                sorting="state.sort"
+                columns="state.columns"
+                infinite-scroll="true"
+                filter="filterQuery"
+                filters="state.filters"
+                data-shared-item
+                data-title="{{opts.savedSearch.lastSavedTitle}}"
+                data-description="{{opts.savedSearch.description}}"
+                minimum-visible-rows="minimumVisibleRows"
+                render-counter
+                on-add-column="addColumn"
+                on-change-sort-order="setSortOrder"
+                on-move-column="moveColumn"
+                on-remove-column="removeColumn"
+              ></doc-table>
+
+              <a tabindex="0" id="discoverBottomMarker"></a>
+
+              <div ng-if="rows.length == opts.sampleSize" class="discover-table-footer">
+                These are the first {{opts.sampleSize}} documents matching
+                your search, refine your search to see others.
+                <a kbn-accessible-click ng-click="scrollToTop()">Back to top.</a>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </main>
+</discover-app>
