+++ public/templates/discover.jade
@@ -1,3 +1,3 @@
 include ./discover.head
-include ./discover.html
+include ./directives/discover-template.html
 include ./footer.foot
